python_requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-swagger==1.0.16
aiosignal==1.2.0
ansible==5.9.0
ansible-core==2.12.6
async-lru==1.0.3
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cisco-gnmi==1.0.15
cisco-mdt==0.0.1
cryptography==37.0.2
dill==0.3.5.1
distro==1.7.0
frozenlist==1.3.0
genie==22.5
genie.libs.clean==22.5
genie.libs.conf==22.5
genie.libs.filetransferutils==22.5
genie.libs.health==22.5
genie.libs.ops==22.5
genie.libs.parser==22.5
genie.libs.sdk==22.5
gitdb==4.0.9
GitPython==3.1.18
grpcio==1.46.3
idna==3.3
Jinja2==3.1.2
jsonpickle==2.2.0
junit-xml==1.9
lxml==4.9.0
MarkupSafe==2.1.1
multidict==6.0.2
ncclient==0.6.15
netaddr==0.8.0
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.4.0
prettytable==3.3.0
protobuf==3.20.1
psutil==5.9.1
pyats==22.5
pyats.aereport==22.5
pyats.aetest==22.5
pyats.async==22.5
pyats.connections==22.5
pyats.datastructures==22.5
pyats.easypy==22.5
pyats.kleenex==22.5
pyats.log==22.5
pyats.reporter==22.5
pyats.results==22.5
pyats.tcl==22.5
pyats.topology==22.5
pyats.utils==22.5
pycparser==2.21
pyftpdlib==1.5.6
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.27.1
resolvelib==0.5.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scp==0.15.0
six==1.16.0
smmap==5.0.0
tftpy==0.8.0
tqdm==4.64.0
unicon==22.5
unicon.plugins==22.5
urllib3==1.26.9
wcwidth==0.2.5
xmltodict==0.13.0
yamllint==1.26.3
yang.connector==22.5.1
yarl==1.7.2
